telegram-webapp/requirements.txt
python-telegram-bot==20.7
flask==2.3.3
flask-socketio==5.3.6